Implement ftoa function in stdlib (#11)
* Ftoa (#7)

* Add files via upload

* double to string conversion

* Update ftoa.c

* Update objects.mk

* ftoa #2

* Delete ftoa.c

* Changed after comments

pow will be implemented in math.h


* Clean ftoa

* Do more ftoa cleanup

* Fix types and add ftoa declaration

* Take ndigit as unsigned

* Update ceil.c

+++ include/math.h
@@ -1,8 +1,8 @@
 #ifndef MATH_H
 #define MATH_H
 
-double ceil(double x);
-double fabs(double x);
+float ceil(float x);
+float fabs(float x);
 float fmax(float x, float y);
 float fmin(float x, float y);
 float pow(float x, int n);
