+++ libc/math/ceil.c
@@ -1,9 +1,9 @@
 #include <math.h>
 
-double ceil(double x) {
+float ceil(float x) {
   int n = (int)x;
   if (n == x) {
     return x;
   }
-  return (double)n + 1;
+  return (float)n + 1;
 }
